add, edit,delete

+++ src/context/user.context.jsx
@@ -3,7 +3,7 @@ import { createContext, useState } from 'react';
 const UserContext = createContext('test');
 
 const DEFAULT_DATA = {
-	setting: { activeUserID: 1 },
+	setting: { activeUserID: 1, activeJurnal: 0 },
 	users: [
 		{
 			id: 0,
@@ -42,6 +42,16 @@ const DEFAULT_DATA = {
 			name: 'Vasile',
 			jurnalItems: [],
 		},
+		{
+			id: 3,
+			name: 'Huilo',
+			jurnalItems: [],
+		},
+		{
+			id: 4,
+			name: 'Aliona',
+			jurnalItems: [],
+		},
 	],
 };
 export const UserContextProvider = ({ children }) => {
@@ -61,18 +71,63 @@ export const UserContextProvider = ({ children }) => {
 	const addJurnal = (newJurnalItem, activeUser) => {
 		setData((data) => {
 			const newData = { ...data };
-			newData.users[activeUser].jurnalItems.push({
-				...newJurnalItem,
-				id: Math.max(...newData.users[activeUser].jurnalItems.map((el) => el.id), 0) + 1,
-			});
-			localStorage.setItem('data', JSON.stringify(newData));
+			if (newJurnalItem.id === 0) {
+				newData.users[activeUser].jurnalItems.push({
+					...newJurnalItem,
+					id: Math.max(...newData.users[activeUser].jurnalItems.map((el) => el.id), 0) + 1,
+				});
+			} else {
+				const elementToReplace = newData.users[activeUser].jurnalItems.find(
+					(el) => el.id === newJurnalItem.id,
+				);
+				const indexToReplace = newData.users[activeUser].jurnalItems.indexOf(elementToReplace);
 
+				if (indexToReplace !== -1) {
+					newData.users[activeUser].jurnalItems.splice(indexToReplace, 1, newJurnalItem);
+				}
+			}
+			localStorage.setItem('data', JSON.stringify(newData));
+			console.log(newData);
 			return newData;
 		});
 	};
 
+	const deleteJurnal = (id) => {
+		if (confirm('esti sigur?'))
+			setData((data) => {
+				const newData = { ...data };
+
+				const elementToReplace = newData.users[newData.setting.activeUserID].jurnalItems.find(
+					(el) => el.id === id,
+				);
+				const indexToReplace =
+					newData.users[newData.setting.activeUserID].jurnalItems.indexOf(elementToReplace);
+
+				if (indexToReplace !== -1) {
+					newData.users[newData.setting.activeUserID].jurnalItems.splice(indexToReplace, 1);
+				}
+
+				localStorage.setItem('data', JSON.stringify(newData));
+				console.log(newData);
+				return newData;
+			});
+		setActiveJurnal(0);
+	};
+	const setActiveJurnal = (id) => {
+		setData((data) => {
+			return {
+				...data,
+				setting: {
+					...data.setting,
+					activeJurnal: +id,
+				},
+			};
+		});
+	};
+
 	return (
-		<UserContext.Provider value={{ data, setData, setActiveUserContext, addJurnal }}>
+		<UserContext.Provider
+			value={{ data, setData, setActiveUserContext, addJurnal, deleteJurnal, setActiveJurnal }}>
 			{children}
 		</UserContext.Provider>
 	);
